feat(game): drop the /play prefix from level urls
